Move funcs to utils module
