+++ src/packages.rs
@@ -1,9 +1,9 @@
+use crate::utils::get_lnreader;
+
 use regex::Regex;
 use std::collections::HashMap;
 use std::fs;
-use std::fs::{DirEntry, File};
-use std::io::{self, BufRead};
-use std::path::Path;
+use std::fs::DirEntry;
 use std::{fmt, path::PathBuf};
 
 const DISTRMETA_NAME_REGEX: &'static str = r"^(?:n|N)ame:(\s)?(?<name>[a-zA-Z0-9._-]+)";
@@ -82,26 +82,6 @@ impl fmt::Display for DistrMeta {
 
 const METADATA_DIR_SUFFIX: &'static str = ".dist-info";
 const METADATA_FILE_NAME: &'static str = "METADATA";
-
-/// from https://doc.rust-lang.org/rust-by-example/std_misc/file/read_lines.html
-fn read_lines<P>(filename: P) -> io::Result<io::Lines<io::BufReader<File>>>
-where
-    P: AsRef<Path>,
-{
-    let file = File::open(filename)?;
-    Ok(io::BufReader::new(file).lines())
-}
-
-fn get_lnreader<P, F>(filename: P, stop_func: F) -> Result<impl Iterator<Item = String>, io::Error>
-where
-    P: AsRef<Path>,
-    F: Fn(&Result<String, std::io::Error>) -> bool,
-{
-    let line_reader = read_lines(&filename)?;
-    Ok(line_reader
-        .take_while(move |line| stop_func(line))
-        .map(|l| l.unwrap()))
-}
 
 /// Get iterator which filter dir entries by metadata suffix
 fn get_meta_dirs(env_path: &PathBuf) -> impl Iterator<Item = DirEntry> {
